refactor(logString) : Added custom AudioLog
Introduced method in folder "AudioScript utilities" to call a custom preformated DebugLog, which indicates in its arguments where it comes from (the C# script it was triggered from), where it is attached (the gameObject getting the method) and additional string for more context if needed.

Replace Debug.Log in "Projectile.cs" by AudioDebugLog.LogAudio.