chore(audio): Poison projectile sound
TODO : refactoring of RTPC update logic
Current system is giving Wwise trouble to update rapidly enough to get
the accurate RTPC value for RTPC_Weapon because the method is called
into the initialize projectile, which is fine for slow to medium-rate
fire projectiles but is an issue for rapid-rate projectiles like the
gatling (it needs a check AND then it posts the event).
Therefore, the gatling is currently posting the previous gun sound and
update itself once you re-trigger fire.